requirements.txt
distorm3
pycrypto
yara-python
pefile
pbkdf2
distorm3
